some little modifications nothign of big importance
@aantn I think the memory leak is in zeitgeist_panel_widgets.py. We need 
to delete the iterators in the Liststore to prevent memory leaks.
I also think that on new event we should not really reload the whole 
view again but only the last day and in it we dont replace the whole day 
but only add the latest entry into the view! this will save us alot of 
memory and should be doable!
cheers