Add a new script to run all of the unit tests at once.
Tests must be in test/ and end with -test.py in order to be run. (The benchmark test was excluded on purpose.) Please note that there's currently something wrong with the logger test. It fails when it is run alone and stops failing when it is run as part of the test suite.