Make the engine-engine-test.py independant of the storm_base module
Add methods close() and is_closed() to the ZeitgeistEngine API. We need this to do proper unit testing without gross hacks.

Make the factory and default engine getter in _zeitgeist/engine/__init__ check if the engine is closed.

With this commit in place you should be able to run the unit tests with another engine by setting the ZEITGEIST_ENGINE environment variable. Fx. setting it to 'querymancer'...