Fix and change DBus naming scheme
The bus name and interface name for the Log was mixed up,
and both where set to org.gnome.zeitgeist.LogManager.

With this commit I changed the bus name of the engine process
to org.gnome.zeitgeist.Engine. And the interface name for our
Logs are now org.gnome.zeitgeist.Log.

Our primary Log object has also changed path to
/org/gnome/zeitgeist/log/activity instead of
/org/gnome/zeitgeist/log.
This gives room for private logs if we want that in the future.