recent.py: make full use of the data-source registry
The GtkRecentlyUsed data-source does now connect to the DataSourceRegistered
signal, in addition to requesting everything at startup, to have a complete
list of actors for which there are loggers.

The corresponding event interpretations are also considered, so that if
CreateEvents are being inserted but VisitEvents not, the data-source will
continue taking care of those.

Remember that in theory the assumption "if a data-source registered an event 
template with actor X and event interpretation Y, so all Y events for app
X are being logged by it" can't be made, but in this particular case of the
GtkRecentlyUsed data-source we'll consider it as valid.

Also note that the changed in this commit don't take into account whether
the data-sources providing logging for a particular actor are currently running
or not. This makes sense, as else we'd end up with repeated events anyway
(when the events aren't logged at exactly the same timestamp by the data-source
and into GtkRecentlyUsed), but given that data-sources can't be deleted from
the registry, ignoring an actor forever is not so nice. The best solution would
probably be adding a RemoveDataSource method to the registry.