Merge ~zeitgeist/zeitgeist/pragma-settings.
We change SQLite's locking mode to EXCLUSIVE, since this gives us a
little reduction in file-system operations and ensures that there is
only one process accessing the database.

We also set the journal mode to the newly available WAL, if the
system has a new enough SQLite (3.7.0+). This results in further
performance improvements but it's important to note that this change
is irreversible.