Add tools/run_fake_zeitgeist.sh
This script sets up a bus in a fake X server (Xvfb) and launches
a Zeitgeist instance (without datahub or FTS) into it.

It then spawns a terminal set up to interact with that instance.
When the terminal is closed (Ctrl+D), the Zeitgeist instance and
bus are terminated.

This is useful for quickly getting a temporary Zeitgeist instance
with a clean database, for testing purposes.