ResultSet: remove next(), seek() and peek()
We could keep next() as deprecated, but we are already breaking plenty
of stuff anyway.

We remove seek() because it doesn't really fit into the iterator
semantics (ie. even if we wrapped a SQL cursor, it wouldn't be easy
to support). As for peek(), we could keep, it's kinda weird to have
it together with a next_value function that'd return the same.