Give the D-Bus service files the correct names
Best-practice is that if you implement a session service whose
well-known name (Name field) is com.example.MyService1, and it
should be service-activatable, then it should install
${datadir}/dbus-1/services/com.example.MyService1.service.
This ensures that two non-conflicting packages cannot provide the
same service name with the same search-path priority (i.e. in the
same directory).

If there are two session services with the same Name in the same
directory, dbus-daemon will arbitrarily choose one of them, which
is unlikely to be the desired result.